JIP-249 FEAT: move BarCodeReader to utils
