+++ src/component/utils/BarcodeReader.js
